ref(events) Remove organization prefix from events views
There is no project level events view so we don't need an organization
prefix to disambiguate the organization views.

Fixes SEN-744
