+++ src/sentry/static/sentry/app/views/events/utils/canIncludePreviousPeriod.jsx
