+++ src/sentry/static/sentry/app/views/events/yAxisSelector.jsx
